More refactoring to prepare for WebAPI scaffolder implementation.
The command line ControllerGeneator class which has Alias attribute is the new file ControllerGen.cs
(I kept the name like that to not conflict with the rename, will rename shortly after a commit...)

The previous ControllerGenerator class is renamed to ControllerGeneratorBase and the functionality
of Empty Controller and Mvc Controller with Context is separated into child classes.
This will help us extend the functionality to Web API Controllers and Mvc Controller with Read/Write Actions.

The command line model is kept such that there is one entry point (dnx gen controller)
for all varieties of controllers and the options determine what gets generated)

Perhaps it's required to show examples in command line help... but for another day...
