+++ src/Microsoft.Extensions.CodeGenerators.Mvc/Controller/ControllerGeneratorModel.cs
@@ -17,7 +17,10 @@ namespace Microsoft.Extensions.CodeGenerators.Mvc
         [Option(Name = "controllerName", ShortName = "name", Description = "Name of the controller")]
         public string ControllerName { get; set; }
 
-        [Option(Name = "restWithNoViews", ShortName = "api", Description = "When specified, generates a Controller with REST style API, noViews is assumed and any view related options are ignored")]
-        public string IsRestController { get; set; }
+        [Option(Name = "restWithNoViews", ShortName = "api", Description = "Specify this switch to generate a Controller with REST style API, noViews is assumed and any view related options are ignored")]
+        public bool IsRestController { get; set; }
+
+        [Option(Name = "readWriteActions", ShortName = "actions", Description = "Specify this switch to generate Controller with read/write actions when a Model class is not used")]
+        public bool GenerateReadWriteActions { get; set; }
     }
 }
