+++ src/Microsoft.Extensions.CodeGenerators.Mvc/Controller/MvcControllerEmpty.cs
@@ -1,0 +1,51 @@
+﻿using System;
+using System.Threading.Tasks;
+using Microsoft.Dnx.Runtime;
+using Microsoft.Extensions.CodeGeneration;
+using Microsoft.Extensions.CodeGeneration.EntityFramework;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Microsoft.Extensions.CodeGenerators.Mvc.Controller
+{
+    public class MvcControllerEmpty : ControllerGeneratorBase
+    {
+        public MvcControllerEmpty(
+            [NotNull]ILibraryManager libraryManager,
+            [NotNull]IApplicationEnvironment environment,
+            [NotNull]IModelTypesLocator modelTypesLocator,
+            [NotNull]IEntityFrameworkService entityFrameworkService,
+            [NotNull]ICodeGeneratorActionsService codeGeneratorActionsService,
+            [NotNull]IServiceProvider serviceProvider,
+            [NotNull]ILogger logger)
+            : base(libraryManager, environment, modelTypesLocator, entityFrameworkService, codeGeneratorActionsService, serviceProvider, logger)
+        {
+        }
+
+        public override async Task Generate(ControllerGeneratorModel controllerGeneratorModel)
+        {
+            if (!string.IsNullOrEmpty(controllerGeneratorModel.ControllerName))
+            {
+                if (!controllerGeneratorModel.ControllerName.EndsWith(Constants.ControllerSuffix, StringComparison.Ordinal))
+                {
+                    controllerGeneratorModel.ControllerName = controllerGeneratorModel.ControllerName + Constants.ControllerSuffix;
+                }
+            }
+            else
+            {
+                throw new ArgumentException("Controller name is required for an Empty Controller");
+            }
+
+            var layoutDependencyInstaller = ActivatorUtilities.CreateInstance<MvcLayoutDependencyInstaller>(ServiceProvider);
+            await layoutDependencyInstaller.Execute();
+
+            var templateModel = new ClassNameModel(className: controllerGeneratorModel.ControllerName, namespaceName: GetControllerNamespace());
+
+            var templateName = "EmptyController.cshtml";
+            var outputPath = ValidateAndGetOutputPath(controllerGeneratorModel);
+            await CodeGeneratorActionsService.AddFileFromTemplateAsync(outputPath, templateName, TemplateFolders, templateModel);
+            Logger.LogMessage("Added Controller : " + outputPath.Substring(ApplicationEnvironment.ApplicationBasePath.Length));
+
+            await layoutDependencyInstaller.InstallDependencies();
+        }
+    }
+}
